+++ sydtest/src/Test/Syd.hs
@@ -147,6 +147,9 @@ module Test.Syd
 
     -- **** Setup functions
 
+    -- ***** Creating setup functions
+    SetupFunc (..),
+
     -- ***** Using setup functions
 
     -- ****** Around
@@ -157,15 +160,6 @@ module Test.Syd
     -- ****** AroundAll
     setupAroundAll,
     setupAroundAllWith,
-
-    -- ***** Creating setup functions
-    SetupFunc (..),
-    makeSimpleSetupFunc,
-    useSimpleSetupFunc,
-    connectSetupFunc,
-    composeSetupFunc,
-    wrapSetupFunc,
-    unwrapSetupFunc,
 
     -- *** Declaring different test settings
     modifyMaxSuccess,
@@ -265,5 +259,9 @@ sydTestWith sets spec = do
   resultForest <- sydTestResult sets spec
   when (shouldExitFail (timedValue resultForest)) (exitWith (ExitFailure 1))
 
+-- | Run a test suite during test suite definition.
+--
+-- This function only exists for backward compatibility.
+-- You can also just use 'liftIO' instead.
 runIO :: IO e -> TestDefM a b e
 runIO = liftIO